requirements.txt
discord.py
pytest
pytest-asyncio
python-dotenv
aiohttp